Avoid ODR violation with `gemm_extra_cols` on PPC
When `EIGEN_ALTIVEC_MMA_DYNAMIC_DISPATCH` is enabled `gemmMMA` will
instantiate `gemm_extra_cols` again (first in `gemm`).
However as `MatrixProductMMA.h` enables `#pragma GCC target("cpu=power10,htm")`
this leads to a different binary version of the function with the same name
which is an ODR violation.
This in turn leads to crashes in `gemm->gemm_extra_cols` when it happens
to use the MMA instantiation.

Solve this by using a new name for the function similar to
`gemmMMA_cols` vs `gemm_cols`.

Fixes #2608