Avoid resursive block definitions.
In algorithms that require recursive calls on blocks (e.g. as required by
tlapack), recursive block class definitions can lead to endless compile
loops, causing the compiler to crash.  Since blocks are contiguous
sections of an underlying expression, we are ble to avoid
this by referring to the original undelrying expression when taking a
block of a block.  This should also end up being more efficient.

Note that this is a breaking change, in that existing code that explicitly
stores block-of-block types will now have a type mismatch.  There were
only a small handful of such cases encountered in the tests.