Add EIGEN_DEVICE_FUNC attribute to run methods
The EIGEN_DEVICE_FUNC attribute was added to the run methods in the Transform class. The purpose of this change is to make the run methods callable from device code when compiling with nvcc, the NVIDIA CUDA compiler. This is beneficial for code that needs to use Eigen functionality on the GPU.