Fix PPC rand and other failures.
The tests previously indirectly forced all random integer values to be in the range [-10, 10]
for tests on PPC by setting a macro.  This broke the new `rand` tests for checking the full
range of values.

Removing that hack, however, caused other failures for tests involving signed integer overflows
(e.g. matrix multiplication of random integer matrices).  Signed int overflow in computations like
this is technically UB.  The results are predictable/consistent on other platforms, but apparently
not on ppc64le.

Fixed the failing PPC tests by limiting the random integer inputs in a handful of select places.
We may still be overflowing signed ints in places, but at least the tests pass.

Also needed to add `ploadquad` for `Packet16(u)c`.  Not sure how the packetmath tests are passing
in the CI, but it's failing when running locally on qemu due to the missing function.