Fix UB in bool packetmath test.
It's UB to explicitly load 0xFF into a bool for the select mask.  Changing the mask
to use valid boolean values causes `pselect` to fail, since the intel blend intrinsic
checks the high bit instead of the low bit.  Modified the test and `pselect` to work
with valid bool masks.