Remove implicit this capture in lambdas
When declaring a lambda with a value-capture default `[=, ...]`, the
`this` pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior (https://reviews.llvm.org/D142639)
and produces a warning in newer versions of clang.

This commit simply makes the implicit captures explicit, preventing
the warning. To maintain compatibility with c++14 and newer, it does
so by explicitly capturing all variables without using a capture-
default.