Fix STL feature detection for c++20.
To use c++20 features in the STL, not only does the compiler need to support c++20 (`__cplusplus >= 202002L`),
the STL library itself needs to as well.

This can break when using a modern compiler (e.g. clang) with an older STL library (e.g. gcc's libstdc++).

We were seeing breakages when compiling JAX/TensorFlow with c++20 for manylinux2014, which relies on libstdc++ v9.