Add "Version" file and update version.
A default "Version" file is added to set the current developer version
to 5.0.0-dev+master.  This file will be updated for individual releases.

Modified cmake install scripts so that the Version file is updated on
install to include the git commit (if installed from git),
e.g. 5.0.0-dev+70d8d99d.