Added versioning for shared libraries.
According to the packaging policies for shared libraries in Linux,
shared libraries should look like this:
libfoo.so.${MAJOR}.${MINOR}.${PATCH}

The following symlinks to this library should also be created:
libfoo.so .${MAJOR} -> libfoo.so .${MAJOR}.${MINOR}.${PATCH}
libfoo.so -> libfoo.so .${MAJOR}.${MINOR}.${PATCH}

In this case, the libfoo library package contains the following files:
The library itself: libfoo.so .${MAJOR}.${MINOR}.${PATCH}
As well as a symlink to it, which contains the library name and the ABI
version.:
libfoo.so.${MAJOR}

The following symlink is placed in the libfoo-devel package:
libfoo.so

This is the standard packaging method for many distributions.
Look:
https://docs.fedoraproject.org/en-US/packaging-guidelines/#_shared_libraries
https://www.debian.org/doc/debian-policy/ch-sharedlibs.html

Currently, when using the cmake options EIGEN_BUILD_BLAS=ON and
EIGEN_BUILD_LAPACK=ON, libraries are being built libeigen_blas.so and
libeigen_lapack.so . They do not have an ABI specific, which is not
suitable for packaging these libraries in distributions without
violating the policy.