Expand CMake compatibility range for single-version specifications.
We initially limited the configuration to only accept versions matching
the exact requested version, so that
```
find_package(Eigen3 5.1)
```
would only match versions starting specifically with 5.1, and not 5.2.
Due to discussions in #3004, we are expanding the range to consider
semantic versioning, so that
```
find_package(Eigen3 5.1)
```
will match 5.1.0 and up to (but not including) 6.0.0.  For the previous
behavior, use the `EXACT` specifier, so
```
find_package(Eigen3 5.1 EXACT)
```
will still match the range 5.1.0 up to (but not including) 5.2.0.

Fixes #3004.